Updated the readme to be more accurate. Also added the following catgory validations: Don't allow for the same category name, make sure category can't be deleted if it is in use, and also fixed a bug with catgory names being empty throwing php errors and not my defined errors.

+++ controllers/category/category_delete.php
@@ -6,6 +6,12 @@ if(isset($_SESSION['uid'])) {
             if (isset($_POST['id'])) {
                 $id = $_POST['id'];
 
+                if (\classes\models\article\Category::isCategoryInUse($id)) {
+                    $_SESSION['in_use'] = "This category is currently in use, please either change or delete the article that the category belongs to.";
+                    header('Location: /admin/categories');
+                    die();
+                }
+
                 if (!\classes\models\article\Category::delete($id)) {
                     $_SESSION['delete_error'] = "There was an error trying to delete this article, please refresh the page and try again.";
                     header('Location: /admin/categories');
